feat: transform mongoose.js to typescript
